% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cand_contrib.R
\name{cand_contrib}
\alias{cand_contrib}
\title{Top contributor to specific candidate}
\usage{
cand_contrib(candidate_id = NULL, cycle = 2018, api_key = get_os_key())
}
\arguments{
\item{candidate_id}{Candidate ID}

\item{cycle}{2012, 2014, 2016, 2018 (blank or out of range cycle will return most recent cycle)}
}
\description{
Returns top contributors to specified candidate for a House or Senate seat or member of Congress. These are 6-year numbers for senators/Senate candidates; 2-year numbers for representatives/House candidates.
}
