% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cand_ind_contrib.R
\name{cand_ind_contrib}
\alias{cand_ind_contrib}
\title{Candidate contributions by specified industry}
\usage{
cand_ind_contrib(candidate_id, industry_id, cycle = 2018,
  api_key = get_os_key())
}
\arguments{
\item{candidate_id}{Candidate ID}

\item{industry_id}{a 3-character industry code}

\item{cycle}{2012, 2014, 2016, 2018 (blank or out of range cycle will return most recent cycle)

#' @examples

\dontrun{
# jeff sessions contributions from real estate
cand_ind_contrib("N00003062", "F10", 2018)
}}
}
\description{
Provides total contributed to specified candidate from specified industry. Senate data reflects 2-year totals.
}
