% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cand_industry.R
\name{cand_industry}
\alias{cand_industry}
\title{Candidate Industry Contributions}
\usage{
cand_industry(candidate_id, cycle = 2018, api_key = get_os_key())
}
\arguments{
\item{candidate_id}{Candidate ID}

\item{cycle}{2012, 2014, 2016, 2018 (blank or out of range cycle will return most recent cycle)}
}
\description{
Provides the top ten industries contributing to a specified candidate for a House or Senate seat or member of Congress. These are 6-year numbers for Senators/Senate candidates; 2-year total for Representatives/House candidates.
}
