% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cand_sector.R
\name{cand_sector2}
\alias{cand_sector2}
\title{get_sector2 Candidate contributions by sector}
\usage{
cand_sector2(candidate_id, cycle = 2018, api_key = get_os_key())
}
\arguments{
\item{candidate_id}{candidate ID}

\item{cycle}{2012, 2014, 2016, 2018; leave blank for latest cycle}
}
\description{
Provides sector total of specified politician's receipt
}
