% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cand_summary.R
\name{cand_summary}
\alias{cand_summary}
\title{Candidate Summary
Provides summary fundraising information for specified politician}
\usage{
cand_summary(candidate_id = NULL, cycle = NULL,
  api_key = get_os_key())
}
\arguments{
\item{candidate_id}{CandidateID}

\item{cycle}{2012, 2014, 2016, 2018; leave blank for latest cycle}
}
\description{
Candidate Summary
Provides summary fundraising information for specified politician
}
