% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/committee_ind.R
\name{committee_ind}
\alias{committee_ind}
\title{Congressional committee donations by industry}
\usage{
committee_ind(committee_id, industry_id, session = 115, api_key = get_os_key())
}
\arguments{
\item{committee_id}{Committee ID in CQ format}

\item{industry_id}{112 (uses 2012 data), 113 (uses 2014 data), 114 (uses 2016 data), or 115 (uses 2018 data); leave blank for latest congress}

\item{session}{112 (uses 2012 data), 113 (uses 2014 data), 114 (uses 2016 data), or 115 (uses 2018 data); leave blank for latest congress}
}
\description{
Provides summary fundraising information for a specific committee, industry and congress number
}
