% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_orgs.R
\name{get_orgs}
\alias{get_orgs}
\title{Get Organizations}
\usage{
get_orgs(organization = NULL, api_key = get_os_key())
}
\description{
Get Organization ID from partial name or partial name of organization requested
}
