% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independent_expend.R
\name{independent_expend}
\alias{independent_expend}
\title{Independent Expenditures}
\usage{
independent_expend(api_key = get_os_key())
}
\arguments{
\item{api_key}{Your OpenSecrets API Key}
}
\description{
Method to access the latest 50 independent expenditure transactions reported. Updated 4 times a day.
}
