% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/org_summary.R
\name{org_summary}
\alias{org_summary}
\title{Organization Sumamry}
\usage{
org_summary(org_id, api_key = get_os_key())
}
\arguments{
\item{api_key}{Your OpenSecrets API Key}

\item{org_id:}{(required) CRP orgid (available via getOrgID method)}
}
\description{
Provides 2018 summary fundraising information for the specified organization id
}
