% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personal_finance.R
\name{personal_finance}
\alias{personal_finance}
\title{Member Personal Finance Data}
\usage{
personal_finance(candidate_id, year = 2016, api_key = get_os_key())
}
\arguments{
\item{candidate_id}{candidate ID}

\item{year}{2013, 2014, 2015 and 2016 data provided where available}

\item{api_key}{Your OpenSecrets API Key}
}
\description{
Returns data on the personal finances of a member of Congress, as well as judicial and executive branches
}
