% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legislators.R
\name{process_legislators}
\alias{process_legislators}
\title{@title process_legislators}
\usage{
process_legislators(res = NULL)
}
\description{
@description helper function to handle response (from API call)
}
\details{
@details  NOT exported
 @param res response from API call
}
