% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legislators.R
\name{get_legislators}
\alias{get_legislators}
\title{Get Legislators}
\usage{
get_legislators(state = NULL, api_key = get_os_key())
}
\description{
Provides a list of 115th Congressional legislators and associated attributes for specified subset (state or specific CID)
}
\details{
@details get_legislators() returns either a tibble with information for
 all legislators for a state, or a list with information about one
 specific legislator.  If `state` has 2 characters, the function searches
 for state.   If `state` has 9 characters, the function searches for
 specific legislator.  If `state` is neither or does not pass pattern
 checks, an error is returned.

 The `res` response from GET API call is slightly different if `arg` is
 state or if `arg` is a legislator.   This requires 2 ways to convert to res
 into the appropriate tibble.  In particular, in case of legislator the
 field names include `@attributes` that must be removed.

 @param state Either (1) 2-letter state code or (2) 9-letter candidate_id
 (cid)
}
\examples{
\dontrun{
     # Bonamici 
       state = "OR"
       candidate_id  <- "N00033474"
       state = candidate_id
       api_key = get_os_key()
       cycle  <- 2018
       get_legislators(arg=candidate_id, api_key = api_key)
       }

}
