% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_legislators.R
\name{get_legislators}
\alias{get_legislators}
\title{Get Legislators}
\usage{
get_legislators(state = NULL, api_key = get_os_key())
}
\description{
Provides a list of 115th Congressional legislators and associated attributes for specified subset (state or specific CID)
}
